requirements.txt
alembic==1.0.0
amqp==2.3.2
aniso8601==3.0.2
billiard==3.5.0.4
celery==4.2.1
certifi==2018.8.24
chardet==3.0.4
click==6.7
dnspython==1.15.0
enum-compat==0.0.2
eventlet==0.24.1
Flask==1.0.2
Flask-Login==0.4.1
flask-marshmallow==0.9.0
Flask-Migrate==2.2.1
Flask-RESTful==0.3.6
Flask-Session==0.3.1
Flask-SocketIO==3.0.2
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Flask-WTF==0.14.2
greenlet==0.4.15
idna==2.7
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.1
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.6
marshmallow-sqlalchemy==0.14.1
monotonic==1.5
psycopg2-binary==2.7.5
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
python-engineio==2.2.0
python-socketio==2.0.0
pytz==2018.5
redis==2.10.6
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.12
urllib3==1.23
vine==1.1.4
Werkzeug==0.14.1
WTForms==2.2.1
